Port value buffering and change function
Added a buffer to capture the port status and a change function to
report if it’s changed or not

Read pins from the buffer using digitalReadBuf(pin)

Use change() inside and if statement to only run code if the port has
changed.

e.g.

onechip.digitalRead() //pull all values in to a buffer
if (onechip.change()) {
//do stuff
}